split up component
