+++ services/src/user/user-service.ts
@@ -1,6 +1,6 @@
 // TODO: Investigate why the services path alias does not work here
 import { compare, hash } from '../crypto/hash'
-import type { ChangePasswordPayload } from '../../../src/routes/(authenticated)/profile/schema'
+import type { ChangePasswordPayload } from '$components/container/profile/schema'
 import { omit } from '../util/omit'
 import type { SelectableUser, User } from './user'
 import { changeUserPasswordById, deleteUserById, getUserById } from './user-repository'
